[#134] Longo tempo para salaver a nova solicitação  https://jira.arinterface.com.br/work_packages/134 front
