+++ .vscode/settings.json
@@ -110,6 +110,7 @@
     "datacriate",
     "datahoje",
     "datalabels",
+    "datanascimento",
     "deescricao",
     "Delet",
     "deleteimg",
@@ -209,6 +210,7 @@
     "instacertificado",
     "instalacao",
     "Instalacao",
+    "Invald",
     "ishidden",
     "Isvalide",
     "iygtasuhvoust",
